requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
bcrypt==3.2.2; python_version >= '3.6'
cffi==1.15.1
click==8.1.3; python_version >= '3.7'
flask-bcrypt==1.0.1
flask==2.2.1
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.1; python_version >= '3.7'
pycparser==2.21
pymysql==1.0.2
werkzeug==2.2.1; python_version >= '3.7'
